camel-imap-store: stalwart compatibility
The Stalwart email server doesn't allow calling SUBSCRIBE to a mail box
that is already subscribed. It looks like that when we first connect to
the server, we don't actually have good state on whether we are
subscribes or not 'si' is always NULL after this call:

    si = camel_lite_store_summary_path((CamelStoreSummary *)store->summary, "INBOX");

Let's ignore the error if the error matches Stalwart's error message.